raspi-signage/files/wait-screen/requirements.txt
-i https://pypi.org/simple
pillow==6.1.0
